Fix additional PyCharm warnings
- data_logger.py: Use string module constants instead of manual character lists
  (string.ascii_letters + string.digits) for validation functions
- data_logger.py: Remove unused imports (Dict, List)
- goggles.py: Rename unused 'frame' parameter to '_frame' in signal handler
- mock_serial_device.py: Remove unused threading import
- staircase.py: Convert np.mean() to float() for type consistency (additional case)

+++ staircase.py
@@ -186,13 +186,13 @@ class StaircaseManager:
 
         if n_reversals == 0 or len(reversals) < n_reversals:
             # Use all reversals
-            threshold = np.mean(reversals)
+            threshold = float(np.mean(reversals))
             logging.info(
                 f"Threshold calculated from all {len(reversals)} reversals: {threshold:.2f}"
             )
         else:
             # Use last n reversals
-            threshold = np.mean(reversals[-n_reversals:])
+            threshold = float(np.mean(reversals[-n_reversals:]))
             logging.info(
                 f"Threshold calculated from last {n_reversals} reversals: {threshold:.2f}"
             )
